refactor(DEV-10629): replace `<MoniteStyleProvider/>` with `<MoniteScopedProviders/>`
