+++ packages/sdk-react/src/index.ts
@@ -1,9 +1,6 @@
 export * from './components';
 
-export {
-  MoniteProvider,
-  MoniteStyleProvider,
-} from './core/context/MoniteProvider';
+export { MoniteProvider } from './core/context/MoniteProvider';
 export * from './core/context/RootElementsProvider';
 export { useMoniteContext } from './core/context/MoniteContext';
 export { toast } from './ui/toast';
@@ -11,3 +8,4 @@ export { toast } from './ui/toast';
 export * from './core/utils';
 export * from './core/hooks';
 export * from './core/queries';
+export { MoniteScopedProviders } from '@/core/context/MoniteScopedProviders';
