+++ packages/sdk-react/src/ui/loadingPage/LoadingPage.tsx
@@ -1,4 +1,4 @@
-import { MoniteStyleProvider } from '@/core/context/MoniteProvider';
+import { MoniteScopedProviders } from '@/core/context/MoniteScopedProviders';
 import { CenteredContentBox } from '@/ui/box';
 import { CircularProgress } from '@mui/material';
